backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
idna==3.10
jsonify==0.5
numpy==2.1.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
urllib3==2.2.3
